commit before instal lib splidejs

+++ database/migrations/2021_03_07_124751_create_tva_restaurants_table.php
@@ -16,7 +16,7 @@ class CreateTvaRestaurantsTable extends Migration
         Schema::create('tva_restaurants', function (Blueprint $table) {
             $table->id();
             $table->string('name_tva')->unique();
-            $table->float('tva');
+            $table->decimal('tva');
             $table->timestamps();
         });
     }
